// src/generated/models/Mdview_mdservicesesModel.ts
export const Mdview_mdservicesesmdview_initiativeareatype = {
  0: 'Service',
  1: 'Service_Process'
} as const;
export type Mdview_mdservicesesmdview_initiativeareatype = keyof typeof Mdview_mdservicesesmdview_initiativeareatype;
export const Mdview_mdservicesesmdview_status = {
  0: 'InProgress',
  1: 'AzureOpenAI'
} as const;
export type Mdview_mdservicesesmdview_status = keyof typeof Mdview_mdservicesesmdview_status;
export const Mdview_mdservicesesmdview_unitofmeasure = {
  0: 'Request',
  1: 'Transaction',
  2: '_264000'
} as const;
export type Mdview_mdservicesesmdview_unitofmeasure = keyof typeof Mdview_mdservicesesmdview_unitofmeasure;
export const Mdview_mdservicesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Mdview_mdservicesesstatecode = keyof typeof Mdview_mdservicesesstatecode;
export const Mdview_mdservicesesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Mdview_mdservicesesstatuscode = keyof typeof Mdview_mdservicesesstatuscode;

export interface Mdview_mdservicesesBase {
  importsequencenumber?: number;
  mdview_actualcostsaving?: number;
  mdview_actualgolivedate?: string;
  mdview_aitechnologyplatform?: string;
  mdview_annualvolume?: string;
  mdview_businessowner?: string;
  mdview_centralinitiativeid?: string;
  mdview_departmentorservicearea?: string;
  mdview_division?: string;
  mdview_eligibility?: string;
  mdview_fullresponsetime?: number;
  mdview_initiativeareadetails?: string;
  mdview_initiativeareatype?: Mdview_mdservicesesmdview_initiativeareatype;
  mdview_initiativedescription?: string;
  mdview_initiativereferencenumber?: string;
  mdview_initiativetitle?: string;
  mdview_mdservicesid: string;
  mdview_modelnameandversion?: string;
  mdview_plannedcompletiondate?: string;
  mdview_productivitygainpercentage?: number;
  mdview_ptuconsumptionovertime?: string;
  mdview_ptuutilizationpercentage?: number;
  mdview_queuewaittime?: number;
  mdview_relatedprocess?: string;
  mdview_requestsperminute?: number;
  mdview_retrycount?: number;
  mdview_sapmodule?: string;
  mdview_servicestate?: string;
  mdview_servicestatus?: string;
  mdview_sourcerecordid?: string;
  mdview_sourcesystem?: string;
  mdview_status?: Mdview_mdservicesesmdview_status;
  mdview_submissiondate?: string;
  mdview_targetcostsaving?: number;
  mdview_targetftesaving?: number;
  mdview_timeconsumedhours?: string;
  mdview_timesaveddescription?: string;
  mdview_tokensperapicall?: number;
  mdview_tokensperminute?: number;
  mdview_unitofmeasure?: Mdview_mdservicesesmdview_unitofmeasure;
  overriddencreatedon?: string;
  statecode: Mdview_mdservicesesstatecode;
  statuscode?: Mdview_mdservicesesstatuscode;
  timezoneruleversionnumber?: number;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Mdview_mdserviceses extends Mdview_mdservicesesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: number;
  mdview_actualcostsaving_base?: number;
  mdview_initiativeareatypename?: string;
  mdview_statusname?: string;
  mdview_targetcostsaving_base?: number;
  mdview_unitofmeasurename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}