// src/generated/models/Mdview_mdagentsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Mdview_mdagentsesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Mdview_mdagentsesstatecode = keyof typeof Mdview_mdagentsesstatecode;
export const Mdview_mdagentsesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Mdview_mdagentsesstatuscode = keyof typeof Mdview_mdagentsesstatuscode;

export interface Mdview_mdagentsesBase {
  importsequencenumber?: number;
  mdview_agentname?: string;
  mdview_initiativeid?: string;
  mdview_initiativetitle?: string;
  mdview_mcpservers?: string;
  mdview_mdagentsid: string;
  mdview_modelsused?: string;
  mdview_ptuusage?: string;
  mdview_sourcefile?: string;
  mdview_systemsintegrated?: string;
  mdview_targetendusers?: string;
  overriddencreatedon?: string;
  statecode: Mdview_mdagentsesstatecode;
  statuscode?: Mdview_mdagentsesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Mdview_mdagentses extends Mdview_mdagentsesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}